hoist navigation.navigate switch to gamesessioncontainer. limit gamesession subscription to when gamecode is entered
