+++ mobile/src/Navigator/index.js
@@ -17,151 +17,154 @@ import StudentName from "../Student/screens/StudentName"
 const Stack = createStackNavigator()
 
 const AppContainer = () => {
-    return (
-        <GameSessionContainer>
-            {({
-                gameSession,
-                fetchGameSessionByCode,
-                team,
-                teamMember,
-                setTeamInfo,
-                teamAvatar,
-                saveTeamAvatar,
-                clearStorage,
-            }) => (
-                <NavigationContainer >
-                    <Stack.Navigator
-                        initialRouteName="JoinGame"
-                        screenOptions={{
-                            headerShown: false,
-                            gestureEnabled: false
-                        }}
-                    >
-                        <Stack.Screen name="JoinGame">
-                            {(props) => (
-                                <JoinGame
-                                    {...props}
-                                    gameSession={gameSession}
-                                    team={team}
-                                    teamMember={teamMember}
-                                    clearStorage={clearStorage}
-                                />
-                            )}
-                        </Stack.Screen>
-                        <Stack.Screen name="EnterGameCode">
-                            {(props) => (
-                                <EnterGameCode
-                                    {...props}
-                                    fetchGameSessionByCode={fetchGameSessionByCode}
-                                />
-                            )}
-                        </Stack.Screen>
-                        <Stack.Screen name="StudentName">
-                            {(props) => (
-                                <StudentName
-                                    {...props}
-                                    gameSession={gameSession}
-                                    setTeamInfo={setTeamInfo}
-                                />
-                            )}
-                        </Stack.Screen>
-                        <Stack.Screen name="SelectTeam">
-                            {(props) => (
-                                <SelectTeam
-                                    {...props}
-                                    gameSession={gameSession}
-                                    team={team}
-                                    teamMember={teamMember}
-                                    saveTeamAvatar={saveTeamAvatar}
-                                />
-                            )}
-                        </Stack.Screen>
-                        <Stack.Screen name="StudentGameIntro">
-                            {(props) => (
-                                <StudentGameIntro
-                                    {...props}
-                                    gameSession={gameSession}
-                                    teamMember={teamMember}
-                                    team={team}
-                                    teamAvatar={teamAvatar}
-                                />
-                            )}
-                        </Stack.Screen>
-                        <Stack.Screen name="PregameCountDown">
-                            {(props) => (
-                                <PregameCountDown
-                                    {...props}
-                                />
-                            )}
-                        </Stack.Screen>
-                        <Stack.Screen name="PhaseOneBasicGamePlay">
-                            {(props) => (
-                                <PhaseOneBasicGamePlay
-                                    {...props}
-                                    gameSession={gameSession}
-                                    teamMember={teamMember}
-                                    team={team}
-                                    teamAvatar={teamAvatar}
-                                />
-                            )}
-                        </Stack.Screen>
-                        <Stack.Screen name="StartPhase">
-                            {(props) => (
-                                <StartPhase
-                                    {...props}
-                                />
-                            )}
-                        </Stack.Screen>
-                        <Stack.Screen name="PhaseTwoBasicGamePlay">
-                            {(props) => (
-                                <PhaseTwoBasicGamePlay
-                                    {...props}
-                                    gameSession={gameSession}
-                                    teamMember={teamMember}
-                                    team={team}
-                                    teamAvatar={teamAvatar}
-                                />
-                            )}
-                        </Stack.Screen>
-                        <Stack.Screen name="PhaseResult">
-                            {(props) => (
-                                <PhaseResult
-                                    {...props}
-                                    gameSession={gameSession}
-                                    teamMember={teamMember}
-                                    team={team}
-                                    teamAvatar={teamAvatar}
-                                    fetchGameSessionByCode={fetchGameSessionByCode}
-                                    setTeamInfo={setTeamInfo}
-                                />
-                            )}
-                        </Stack.Screen>
-                        <Stack.Screen name="ScorePage">
-                            {(props) => (
-                                <ScorePage
-                                    {...props}
-                                    gameSession={gameSession}
-                                    teamMember={teamMember}
-                                    team={team}
-                                    teamAvatar={teamAvatar}
-                                />
-                            )}
-                        </Stack.Screen>
-                        <Stack.Screen name="Leaderboard">
-                            {(props) => (
-                                <Leaderboard
-                                    {...props}
-                                    gameSession={gameSession}
-                                    teamMember={teamMember}
-                                    team={team}
-                                    teamAvatar={teamAvatar}
-                                />
-                            )}
-                        </Stack.Screen>
-                    </Stack.Navigator>
-                </NavigationContainer>
-            )}
-        </GameSessionContainer>
+  return (
+    <NavigationContainer >
+      <GameSessionContainer>
+          {({
+              gameSession,
+              fetchGameSessionByCode,
+              team,
+              teamMember,
+              setTeamInfo,
+              teamAvatar,
+              saveTeamAvatar,
+              clearStorage,
+              subscribeToGame
+          }) => (
+                  <Stack.Navigator
+                      initialRouteName="JoinGame"
+                      screenOptions={{
+                          headerShown: false,
+                          gestureEnabled: false
+                      }}
+                  >
+                      <Stack.Screen name="JoinGame">
+                          {(props) => (
+                              <JoinGame
+                                  {...props}
+                                  gameSession={gameSession}
+                                  team={team}
+                                  teamMember={teamMember}
+                                  clearStorage={clearStorage}
+                              />
+                          )}
+                      </Stack.Screen>
+                      <Stack.Screen name="EnterGameCode">
+                          {(props) => (
+                              <EnterGameCode
+                                  {...props}
+                                  fetchGameSessionByCode={fetchGameSessionByCode}
+                                  subscribeToGame={subscribeToGame}
+                              />
+                          )}
+                      </Stack.Screen>
+                      <Stack.Screen name="StudentName">
+                          {(props) => (
+                              <StudentName
+                                  {...props}
+                                  gameSession={gameSession}
+                                  setTeamInfo={setTeamInfo}
+                              />
+                          )}
+                      </Stack.Screen>
+                      <Stack.Screen name="SelectTeam">
+                          {(props) => (
+                              <SelectTeam
+                                  {...props}
+                                  gameSession={gameSession}
+                                  team={team}
+                                  teamMember={teamMember}
+                                  saveTeamAvatar={saveTeamAvatar}
+                              />
+                          )}
+                      </Stack.Screen>
+                      <Stack.Screen name="StudentGameIntro">
+                          {(props) => (
+                              <StudentGameIntro
+                                  {...props}
+                                  gameSession={gameSession}
+                                  teamMember={teamMember}
+                                  team={team}
+                                  teamAvatar={teamAvatar}
+                              />
+                          )}
+                      </Stack.Screen>
+                      <Stack.Screen name="PregameCountDown">
+                          {(props) => (
+                              <PregameCountDown
+                                  {...props}
+                              />
+                          )}
+                      </Stack.Screen>
+                      <Stack.Screen name="PhaseOneBasicGamePlay">
+                          {(props) => (
+                              <PhaseOneBasicGamePlay
+                                  {...props}
+                                  gameSession={gameSession}
+                                  teamMember={teamMember}
+                                  team={team}
+                                  teamAvatar={teamAvatar}
+                              />
+                          )}
+                      </Stack.Screen>
+                      <Stack.Screen name="StartPhase">
+                          {(props) => (
+                              <StartPhase
+                                  {...props}
+                              />
+                          )}
+                      </Stack.Screen>
+                      <Stack.Screen name="PhaseTwoBasicGamePlay">
+                          {(props) => (
+                              <PhaseTwoBasicGamePlay
+                                  {...props}
+                                  gameSession={gameSession}
+                                  teamMember={teamMember}
+                                  team={team}
+                                  teamAvatar={teamAvatar}
+                              />
+                          )}
+                      </Stack.Screen>
+                      <Stack.Screen name="PhaseResult">
+                          {(props) => (
+                              <PhaseResult
+                                  {...props}
+                                  gameSession={gameSession}
+                                  teamMember={teamMember}
+                                  team={team}
+                                  teamAvatar={teamAvatar}
+                                  fetchGameSessionByCode={fetchGameSessionByCode}
+                                  setTeamInfo={setTeamInfo}
+                              />
+                          )}
+                      </Stack.Screen>
+                      <Stack.Screen name="ScorePage">
+                          {(props) => (
+                              <ScorePage
+                                  {...props}
+                                  gameSession={gameSession}
+                                  teamMember={teamMember}
+                                  team={team}
+                                  teamAvatar={teamAvatar}
+                              />
+                          )}
+                      </Stack.Screen>
+                      <Stack.Screen name="Leaderboard">
+                          {(props) => (
+                              <Leaderboard
+                                  {...props}
+                                  gameSession={gameSession}
+                                  teamMember={teamMember}
+                                  team={team}
+                                  teamAvatar={teamAvatar}
+                              />
+                          )}
+                      </Stack.Screen>
+                  </Stack.Navigator>
+         
+          )}
+      </GameSessionContainer>
+      </NavigationContainer>
     )
 }
